+++ test/tree_test.dart
@@ -51,7 +51,7 @@ void main() {
             Tree(
               key: 'invalid-child-key',
               parent: root,
-              value: ExampleData.example(),
+              data: ExampleData.example(),
             );
           } catch (e) {
             messages = [(e as dynamic).message as String];
@@ -82,7 +82,7 @@ void main() {
             Tree(
               key: '_forbiddenChildKey',
               parent: root,
-              value: ExampleData.example(),
+              data: ExampleData.example(),
             );
           } catch (e) {
             messages = [(e as dynamic).message as String];
@@ -101,7 +101,7 @@ void main() {
           // Act
           final newRoot = Tree<ExampleData>.root(
             key: 'newRoot',
-            value: ExampleData.example(),
+            data: ExampleData.example(),
             children: [child, grandchild],
           );
 
@@ -120,14 +120,14 @@ void main() {
       });
     });
 
-    group('value', () {
+    group('data', () {
       test('returns the value', () {
-        expect(root.value.toJson(), {
+        expect(root.data.toJson(), {
           'me': 'root',
           'hiRoot': 'Hi from root.',
           'isAncestor': true,
         });
-        expect(child.value.toJson(), {
+        expect(child.data.toJson(), {
           'me': 'child',
           'hiChild': 'Hi from child.',
           'isAncestor': false,
@@ -140,7 +140,7 @@ void main() {
             'null': null,
           },
         });
-        expect(grandchild.value.toJson(), {
+        expect(grandchild.data.toJson(), {
           'me': 'grandchild',
           'hiGrandchild': 'Hi from grandchild.',
           'isAncestor': false,
@@ -153,6 +153,34 @@ void main() {
             'null': null,
           },
         });
+      });
+    });
+
+    group('pathSegments', () {
+      test('returns the path segments', () {
+        expect(root.pathSegments, isEmpty);
+        expect(grandpa.pathSegments, ['grandpa']);
+        expect(dad.pathSegments, ['grandpa', 'dad']);
+        expect(me.pathSegments, ['grandpa', 'dad', 'me']);
+        expect(child.pathSegments, ['grandpa', 'dad', 'me', 'child']);
+        expect(grandchild.pathSegments, [
+          'grandpa',
+          'dad',
+          'me',
+          'child',
+          'grandchild',
+        ]);
+      });
+    });
+
+    group('path', () {
+      test('returns the path as string', () {
+        expect(root.path, '/');
+        expect(grandpa.path, '/grandpa');
+        expect(dad.path, '/grandpa/dad');
+        expect(me.path, '/grandpa/dad/me');
+        expect(child.path, '/grandpa/dad/me/child');
+        expect(grandchild.path, '/grandpa/dad/me/child/grandchild');
       });
     });
 
@@ -502,19 +530,19 @@ void main() {
     group('set(key, value)', () {
       test('writes data into the tree', () {
         me.set('../#', {'hello': 'world'});
-        expect(dad.value.data, {'hello': 'world'});
+        expect(dad.data.data, {'hello': 'world'});
 
         me.set('./#', {'new': 'object'});
-        expect(me.value.data, {'new': 'object'});
+        expect(me.data.data, {'new': 'object'});
 
         me.set('./#/new', 'newValue');
-        expect(me.value.data, {'new': 'newValue'});
+        expect(me.data.data, {'new': 'newValue'});
 
         me.set('./#/num', 10, extend: true);
-        expect(me.value.data, {'new': 'newValue', 'num': 10});
+        expect(me.data.data, {'new': 'newValue', 'num': 10});
 
         me.set('./#/num', 11);
-        expect(me.value.data, {'new': 'newValue', 'num': 11});
+        expect(me.data.data, {'new': 'newValue', 'num': 11});
 
         me.set('/#/num', 11, extend: true);
         expect(root.get<int>('#num'), 11);
@@ -631,7 +659,7 @@ void main() {
 
         // Check that the structure is identical
         expect(copy.key, root.key);
-        expect(copy.value.toJson(), root.value.toJson());
+        expect(copy.data.toJson(), root.data.toJson());
         expect(copy.children.length, root.children.length);
 
         final originalChild = root.children.first;
@@ -639,7 +667,7 @@ void main() {
 
         expect(copiedChild, isNot(same(originalChild)));
         expect(copiedChild.key, originalChild.key);
-        expect(copiedChild.value.toJson(), originalChild.value.toJson());
+        expect(copiedChild.data.toJson(), originalChild.data.toJson());
         expect(copiedChild.parent, same(copy));
       });
     });
@@ -659,21 +687,35 @@ void main() {
       });
     });
 
-    group('allNodes', () {
-      test('returns all nodes in the tree', () {
-        expect(root.allNodes(), [root, grandpa, dad, me, child, grandchild]);
-        expect(me.allNodes(), [me, child, grandchild]);
+    group('lsWhere', () {
+      test('returns all paths matching the given condition', () {
+        expect(root.lsWhere((node) => node.key.startsWith('g')), [
+          '/grandpa',
+          '/grandpa/dad/me/child/grandchild',
+        ]);
+
+        expect(me.lsWhere((node) => node.key.contains('h')), [
+          '/child',
+          '/child/grandchild',
+        ]);
       });
     });
 
-    group('allNodesWhere', () {
+    group('lsNodes', () {
+      test('returns all nodes in the tree', () {
+        expect(root.lsNodes(), [root, grandpa, dad, me, child, grandchild]);
+        expect(me.lsNodes(), [me, child, grandchild]);
+      });
+    });
+
+    group('lsNodesWhere', () {
       test('returns all nodes matching the given condition', () {
-        expect(root.allNodesWhere(where: (node) => node.key.startsWith('g')), [
+        expect(root.lsNodesWhere((node) => node.key.startsWith('g')), [
           grandpa,
           grandchild,
         ]);
 
-        expect(me.allNodesWhere(where: (node) => node.key.contains('h')), [
+        expect(me.lsNodesWhere((node) => node.key.contains('h')), [
           child,
           grandchild,
         ]);
@@ -691,12 +733,12 @@ void main() {
       setUp(() {
         parent = Tree<ExampleData>.root(
           key: 'parent',
-          value: ExampleData.example(),
+          data: ExampleData.example(),
         );
 
-        child0 = Tree<ExampleData>.root(key: 'child', value: ev);
-        child1 = Tree<ExampleData>.root(key: 'child', value: ev);
-        child2 = Tree<ExampleData>.root(key: 'child', value: ev);
+        child0 = Tree<ExampleData>.root(key: 'child', data: ev);
+        child1 = Tree<ExampleData>.root(key: 'child', data: ev);
+        child2 = Tree<ExampleData>.root(key: 'child', data: ev);
       });
 
       group('auto correct when nodes have not unique names', () {
@@ -704,7 +746,7 @@ void main() {
           test('constructor', () {
             parent = Tree<ExampleData>.root(
               key: 'name',
-              value: ev,
+              data: ev,
               children: [child0, child1, child2],
             );
             expect(parent.children.elementAt(0).key, 'child0');
@@ -740,9 +782,9 @@ void main() {
         });
 
         test('sibling nodes with same name are renamed', () {
-          final parent = Tree<ExampleData>.root(key: 'parent', value: ev);
-          final child1 = Tree<ExampleData>.root(key: 'child', value: ev);
-          final child2 = Tree<ExampleData>.root(key: 'child', value: ev);
+          final parent = Tree<ExampleData>.root(key: 'parent', data: ev);
+          final child1 = Tree<ExampleData>.root(key: 'child', data: ev);
+          final child2 = Tree<ExampleData>.root(key: 'child', data: ev);
 
           parent.addChildren([child1, child2]);
 
@@ -751,10 +793,10 @@ void main() {
         });
 
         test('multiple sibling nodes with same name are renamed', () {
-          final parent = Tree<ExampleData>.root(key: 'parent', value: ev);
-          final child1 = Tree<ExampleData>.root(key: 'child', value: ev);
-          final child2 = Tree<ExampleData>.root(key: 'child', value: ev);
-          final child3 = Tree<ExampleData>.root(key: 'child', value: ev);
+          final parent = Tree<ExampleData>.root(key: 'parent', data: ev);
+          final child1 = Tree<ExampleData>.root(key: 'child', data: ev);
+          final child2 = Tree<ExampleData>.root(key: 'child', data: ev);
+          final child3 = Tree<ExampleData>.root(key: 'child', data: ev);
 
           parent.addChildren([child1, child2, child3]);